error on design changes in color i will revert commit before this one
